Harden Mistral OCR response hydration and request normalization
- Persist detected document type into the OCR request payload so type-less
  requests still send the required document.type discriminator
- Accept integer-valued confidence scores by validating numeric and casting
  to float in ConfidenceScore and ConfidenceScores
- Use defensive null-coalescing for nullable Image coordinates and optional
  Page images/dimensions to avoid undefined-key warnings
- Add regression tests covering the above

+++ packages/mistral/src/Responses/Ocr/ConfidenceScore.php
@@ -34,12 +34,12 @@ final readonly class ConfidenceScore
         $startIndex = $attributes['start_index'];
 
         Assert::string($text);
-        Assert::float($confidence);
+        Assert::numeric($confidence);
         Assert::integer($startIndex);
 
         return new self(
             $text,
-            $confidence,
+            (float) $confidence,
             $startIndex,
         );
     }
